+++ modules/app-control/src/color.js
@@ -1,5 +1,18 @@
 import hsvToRgb from "hsv-rgb";
 
+const cache = {};
+
+const HSV = /^hsv\((.*)\)$/;
+export function rgb(color) {
+  if (cache[color]) return cache[color];
+  const parsed = HSV.exec(color)[1]
+    .split(",")
+    .map(i => parseInt(i));
+  const c = hsvToRgb.apply(null, parsed);
+  cache[color] = "rgb(" + c[0] + "," + c[1] + "," + c[2] + ")";
+  return cache[color];
+}
+
 const S = [0.5 * 100, 1 * 100];
 const V = [0.9 * 100, 1 * 100];
 const H = [
@@ -20,9 +33,11 @@ function random(range) {
 }
 
 function getColor(number) {
-  const h = H[number % MAX];
-  const c = hsvToRgb(random(h), random(S), random(V));
-  console.log("color", c);
+  const h = random(H[number % MAX]);
+  const s = random(S);
+  const v = random(V);
+  console.log("hsv(" + h + "," + s + "," + v + ")");
+  const c = hsvToRgb(h, s, v);
   return "rgb(" + c[0] + "," + c[1] + "," + c[2] + ")";
 }