+++ modules/audio/src/index.js
@@ -18,12 +18,20 @@ function init(set, events) {
     player.stopAll();
   });
 
-  events.on("/sample/start", function(name) {
-    player.play(name, samples[name], config.samples);
+  events.on("/audio/stop-all", function() {
+    player.names().forEach(function(name) {
+      events.emit("/audio/stop", name);
+    });
   });
 
-  events.on("/sample/stop", function(name) {
+  events.on("/audio/start", function(name) {
+    player.play(name, samples[name], config.samples);
+    events.emit("/audio/started", name, ctx.currentTime);
+  });
+
+  events.on("/audio/stop", function(name) {
     player.stop(name);
+    events.emit("/audio/stoped", name, ctx.currentTime);
   });
 }